+++ coding-frontend/src/page/myPage/OrderList.jsx
@@ -1,30 +1,9 @@
-import {
-    Badge,
-    Box,
-    Button,
-    Divider,
-    Flex,
-    Heading,
-    Input,
-    Modal,
-    ModalBody,
-    ModalContent,
-    ModalFooter,
-    ModalHeader,
-    ModalOverlay,
-    Spinner,
-    Text,
-    Textarea,
-    useDisclosure,
-    VStack
-} from "@chakra-ui/react";
+import {Badge, Box, Button, Divider, Flex, Heading, Spinner, Text, useDisclosure, VStack} from "@chakra-ui/react";
 import React, {useContext, useEffect, useState} from "react";
 import {LoginContext} from "../../component/LoginProvider.jsx";
 import axios from "axios";
 import {useNavigate} from "react-router-dom";
-import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {faStar as emptyStar} from "@fortawesome/free-regular-svg-icons";
-import {faStar as fullStar} from "@fortawesome/free-solid-svg-icons";
+import {ReviewModal} from "./ReviewModal.jsx";
 
 export function OrderList() {
     const account = useContext(LoginContext);
@@ -94,33 +73,11 @@ export function OrderList() {
         return <Spinner/>;
     }
 
-    function handleReview() {
-        const formData = new FormData();
-        formData.append('restaurantId', selectedRestaurant);
-        formData.append('userId', userId);
-        formData.append('rating', rating);
-        formData.append('content', content);
-
-        for (let i = 0; i < files.length; i++) {
-            formData.append('files', files[i]);
-        }
-
-        axios.post('/api/reviews', formData)
-            .then((res) => {
-                console.log("리뷰 저장");
-                onClose();
-            })
-            .catch((error) => console.log("리뷰 저장 실패", error));
-    }
-
     function handleOpenModal(restaurantId) {
         setSelectedRestaurant(restaurantId);
         onOpen();
     }
 
-    function handleStar(num) {
-        setRating(num);
-    }
 
     return (
         <Box maxW="800px" margin="auto" p={5}>
@@ -171,35 +128,13 @@ export function OrderList() {
                         </Flex>
                     </Box>
                 ))}
-                <Modal isOpen={isOpen} onClose={onClose}>
-                    <ModalOverlay/>
-                    <ModalContent>
-                        <ModalHeader>{selectedRestaurant} 리뷰 작성</ModalHeader>
-                        <ModalBody>
-                            {[1, 2, 3, 4, 5].map((num) => (
-                                <Box onClick={() => handleStar(num)} key={num} display="inline-block"
-                                     cursor="pointer">
-                                    <FontAwesomeIcon icon={num <= rating ? fullStar : emptyStar}
-                                                     style={{color: "#FFD43B",}}/>
-                                </Box>
-                            ))}
-                            <Input multiple type={"file"} accept={"image/*"}
-                                   onChange={(e) => setFiles(Array.from(e.target.files))}/>
-                            <Textarea resize={"none"}
-                                      height={"100px"}
-                                      borderColor="gray.400"
-                                      value={content}
-                                      onChange={(e) => setContent(e.target.value)}
-                                      placeholder="음식의 맛, 양, 포장 상태 등 음식에 대한 솔직한 리뷰를 남겨주세요.(선택)"/>
-                        </ModalBody>
-                        <ModalFooter>
-                            <Button colorScheme="blue" onClick={handleReview}>
-                                완료
-                            </Button>
-                            <Button onClick={onClose}>취소</Button>
-                        </ModalFooter>
-                    </ModalContent>
-                </Modal>
+                <ReviewModal
+                    isOpen={isOpen}
+                    onClose={onClose}
+                    selectedRestaurant={selectedRestaurant}
+                    userId={userId}
+                />
+
             </VStack>
         </Box>
     );
